Player Contol Updates

+++ BuildAGame2/src/com/tutorial/main/GameObject.java
@@ -5,11 +5,11 @@ import java.awt.Rectangle;
 
 public abstract class GameObject {
 
-	protected int x, y;
+	protected float x, y;
 	protected ID id;
-	protected int velX, velY;
+	protected float velX, velY;
 	
-	public GameObject(int x, int y, ID id){
+	public GameObject(float x, float y, ID id){
 		this.x = x;
 		this.y = y;
 		this.id = id;
@@ -27,11 +27,11 @@ public abstract class GameObject {
 		this.y = y;
 	}
 	
-	public int getX(){
+	public float getX(){
 		return x;
 	}
 	
-	public int getY(){
+	public float getY(){
 		return y;
 	}
 	
@@ -43,19 +43,19 @@ public abstract class GameObject {
 		return id;
 	}
 	
-	public void setVelX(int velX){
+	public void setVelX(float velX){
 		this.velX = velX;
 	}
 	
-	public void setVelY(int velY){
+	public void setVelY(float velY){
 		this.velY = velY;
 	}
 	
-	public int getVelY(){
+	public float getVelY(){
 		return velY;
 	}
 	
-	public int getVelX(){
+	public float getVelX(){
 		return velX;
 	}
